Patch 2
Добавлено:
*У оружия есть параметр stun. При выстреле пуля замедляет врага в 2 раза на  stun секунд.

Изменено:
*Поменял название переменных и почистил лишнее (типо [Ser...] и public вместе не пишуться если переменная public она ужи сериализована).

+++ Assets/Scripts/player/PlayerHealth.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 
 public class PlayerHealth : MonoBehaviour
 {
-    [SerializeField] private int health;
+    [SerializeField] private float health;
     [SerializeField] private float timeToDamage;
     [SerializeField] private GameObject lozeMenu;
     [SerializeField] private AudioSource audioHit;
